requirements.txt
annotated-types==0.6.0
anyio==4.3.0
azure-ai-documentintelligence==1.0.0b3
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.30.1
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
install==1.3.5
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
msrest==0.7.1
oauthlib==3.2.2
openai==1.25.1
pinecone-client==4.0.0
pydantic==2.7.1
pydantic_core==2.18.2
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
